requirements.txt
Django >= 1.3
django-auth-ldap >= 1.2.0
python-ldap >= 2.0
